tests/wpt/web-platform-tests/html/semantics/embedded-content/the-canvas-element/fallback.basic.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: fallback.basic</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<body class="show_output">

<h1>fallback.basic</h1>
<p class="desc">Fallback content is inserted into the DOM</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Fallback content is inserted into the DOM");
_addTest(function(canvas, ctx) {

_assertSame(canvas.childNodes.length, 1, "canvas.childNodes.length", "1");


});
</script>
